implement comment edit and delete functionality(partly)

+++ src/app/api/post/remove-like/route.ts
@@ -1,15 +1,22 @@
+import { auth } from "@clerk/nextjs/server";
 import { revalidatePath } from "next/cache";
 import { NextResponse, type NextRequest } from "next/server";
 
 import { ROUTES } from "@/utils/routes";
-import { handleError } from "@/lib/error";
+import { ERROR_CODES, ERROR_MESSAGES, handleError } from "@/lib/error";
 import prismaClient from "@/lib/prisma";
 import { LikeRequestSchema } from "@/services/types";
 
 export const DELETE = async (req: NextRequest) => {
+  const user = auth();
+
   try {
     const body = await req.json();
     const { postId, userId } = LikeRequestSchema.parse(body);
+
+    if (user.userId !== userId) {
+      return handleError(ERROR_MESSAGES[ERROR_CODES.USER_IS_NOT_AUTHENTICATED]);
+    }
 
     await prismaClient.likes.deleteMany({
       where: {
